# contracts/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Egypt Industrial Initiatives Platform API",
  "version": "1.1.0"
 },
 "servers": [
  {
   "url": "http://localhost:4000"
  },
  {
   "url": "https://api.industry.gov.eg"
  }
 ],
 "paths": {
...
  "/api/v1/chat/conversations": {
   "get": {
    "summary": "Directory of every other eligible organization + the pair conversation summary",
    "tags": [
     "chat"
    ],
    "responses": {
     "200": {
      "description": "ok"
     },
     "401": {
      "description": "unauthorized"
     },
     "403": {
      "description": "forbidden (factory owners)"
     }
    }
   }
  },
  "/api/v1/chat/unread": {
   "get": {
    "summary": "Unread chat summary for the caller side (bell / badges)",
    "tags": [
     "chat"
    ],
    "responses": {
     "200": {
      "description": "ok"
     },
     "403": {
      "description": "forbidden"
     }
    }
   }
  },
  "/api/v1/chat/conversations/{orgId}/messages": {
   "get": {
    "summary": "List messages of the conversation with peer organization {orgId} (cursor: before / after, limit<=50)",
    "tags": [
     "chat"
    ],
    "parameters": [
     {
      "name": "orgId",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "before",
      "in": "query",
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "after",
      "in": "query",
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "schema": {
       "type": "integer",
       "maximum": 50
      }
     }
    ],
    "responses": {
     "200": {
      "description": "ok"
     },
     "403": {
      "description": "forbidden"
     },
     "404": {
      "description": "org not found / not eligible"
     }
    }
   },
   "post": {
    "summary": "Send to peer organization {orgId} (multipart body + files[] <=5, or JSON {body}); auditor → 403 READ_ONLY",
    "tags": [
     "chat"
    ],
    "parameters": [
     {
      "name": "orgId",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "type": "object",
        "properties": {
         "body": {
          "type": "string",
          "maxLength": 4000
         },
         "files": {
          "type": "array",
          "maxItems": 5,
          "items": {
           "type": "string",
           "format": "binary"
          }
         }
        }
       }
      },
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "body": {
          "type": "string",
          "maxLength": 4000
         }
        }
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "created"
     },
     "400": {
      "description": "validation error / invalid file type / file too large"
     },
     "403": {
      "description": "forbidden"
     },
     "404": {
      "description": "org not found / not eligible"
     },
     "429": {
      "description": "rate limited"
     }
    }
   }
  },
  "/api/v1/chat/conversations/{orgId}/read": {
   "post": {
    "summary": "Mark conversation read for the caller side (re-arms entity email reminder)",
    "tags": [
     "chat"
    ],
    "parameters": [
     {
      "name": "orgId",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "ok"
     },
     "403": {
      "description": "forbidden"
     },
     "404": {
      "description": "not found"
     }
    }
   }
  },
  "/api/v1/chat/attachments/{id}": {
   "get": {
    "summary": "Download chat attachment (inline=1 for images/PDF)",
    "tags": [
     "chat"
    ],
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "inline",
      "in": "query",
      "schema": {
       "type": "string",
       "enum": [
        "1"
       ]
      }
     }
    ],
    "responses": {
     "200": {
      "description": "file"
     },
     "403": {
      "description": "forbidden"
     },
     "404": {
      "description": "not found"
     }
    }
   }
  },
  "/api/v1/chat/oversight": {
   "get": {
    "summary": "Conversations between OTHER organizations (officials / auditor, read-only)",
    "tags": [
     "chat"
    ],
    "responses": {
     "200": {
      "description": "ok"
     },
     "403": {
      "description": "forbidden"
     }
    }
   }
  },
  "/api/v1/chat/oversight/{conversationId}/messages": {
   "get": {
    "summary": "Read a conversation between two other organizations (read-only)",
    "tags": [
     "chat"
    ],
    "parameters": [
     {
      "name": "conversationId",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "ok"
     },
     "403": {
      "description": "forbidden"
     },
     "404": {
      "description": "not found"
     }
    }
   }
  },
...
 }
}